+++ toolkit/scaffold/project/python/example.py
@@ -1,5 +1,4 @@
 from toolkit.config.context import USER_INPUT_CONTEXT
-
 from toolkit.scaffold.project.command import generate_create_project_command
 from toolkit.scaffold.project.template import TEMPLATE_PYTHON_PATH