+++ HISTORY.md
@@ -3,6 +3,7 @@
 v1.5.2 (2022-08-13)
 
 - Improved Qt5 template.
+- Improved Python example template.
 
 v1.5.1 (2022-08-12)