+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+v1.5.3 (2022-08-15)
+
+- Added
+    - golang web example templates
+
 v1.5.2 (2022-08-13)
 
 - Improved Qt5 template.